* Updated the ChilliSource windows project to latest format. * Fixed a number of new warnings and deprecation issues when building with VC++14. * The standardise path methods now allow a double // at the start of the path. This solves issues accessing files access a network on windows.

+++ Source/CSBackend/Platform/Windows/SFML/Base/SFMLWindow.h
@@ -333,7 +333,7 @@ namespace CSBackend
 			sf::ContextSettings m_contextSettings;
 
 			u32 m_preferredRGBADepth = 32;
-			f32 m_preferredFPS = 0.0f;
+			u32 m_preferredFPS = 0;
 
 			bool m_isRunning = true;
 			bool m_isFocused = true;
